--- sbi/neural_nets/build_nets/__init__.py
@@ -1,28 +1,0 @@
-from sbi.neural_nets.build_nets.categorial import build_categoricalmassestimator
-from sbi.neural_nets.build_nets.classifier import (
-    build_linear_classifier,
-    build_mlp_classifier,
-    build_resnet_classifier,
-)
-from sbi.neural_nets.build_nets.flow import (
-    build_made,
-    build_maf,
-    build_maf_rqs,
-    build_nsf,
-    build_zuko_bpf,
-    build_zuko_gf,
-    build_zuko_maf,
-    build_zuko_naf,
-    build_zuko_ncsf,
-    build_zuko_nice,
-    build_zuko_nsf,
-    build_zuko_sospf,
-    build_zuko_unaf,
-)
-from sbi.neural_nets.build_nets.flowmatching_nets import (
-    build_mlp_flowmatcher,
-    build_resnet_flowmatcher,
-)
-from sbi.neural_nets.build_nets.mdn import build_mdn
-from sbi.neural_nets.build_nets.mnle import build_mnle
-from sbi.neural_nets.build_nets.score_nets import build_score_estimator
