+++ sbi/neural_nets/net_builders/__init__.py
@@ -1,0 +1,28 @@
+from sbi.neural_nets.net_builders.categorial import build_categoricalmassestimator
+from sbi.neural_nets.net_builders.classifier import (
+    build_linear_classifier,
+    build_mlp_classifier,
+    build_resnet_classifier,
+)
+from sbi.neural_nets.net_builders.flow import (
+    build_made,
+    build_maf,
+    build_maf_rqs,
+    build_nsf,
+    build_zuko_bpf,
+    build_zuko_gf,
+    build_zuko_maf,
+    build_zuko_naf,
+    build_zuko_ncsf,
+    build_zuko_nice,
+    build_zuko_nsf,
+    build_zuko_sospf,
+    build_zuko_unaf,
+)
+from sbi.neural_nets.net_builders.flowmatching_nets import (
+    build_mlp_flowmatcher,
+    build_resnet_flowmatcher,
+)
+from sbi.neural_nets.net_builders.mdn import build_mdn
+from sbi.neural_nets.net_builders.mnle import build_mnle
+from sbi.neural_nets.net_builders.score_nets import build_score_estimator
