+++ sbi/neural_nets/net_builders/categorial.py
