+++ tests/density_estimator_test.py
@@ -14,10 +14,15 @@ from sbi.neural_nets.embedding_nets import CNNEmbedding
 from sbi.neural_nets.estimators.shape_handling import (
     reshape_to_sample_batch_event,
 )
-from sbi.neural_nets.build_nets import (
+from sbi.neural_nets.net_builders import (
+    build_categoricalmassestimator,
     build_maf,
     build_maf_rqs,
+    build_mdn,
+    build_mlp_flowmatcher,
+    build_mnle,
     build_nsf,
+    build_resnet_flowmatcher,
     build_zuko_bpf,
     build_zuko_gf,
     build_zuko_maf,
@@ -27,10 +32,6 @@ from sbi.neural_nets.build_nets import (
     build_zuko_nsf,
     build_zuko_sospf,
     build_zuko_unaf,
-    build_mnle, build_categoricalmassestimator,
-    build_mlp_flowmatcher,
-    build_resnet_flowmatcher,
-    build_mdn,
 )
 
 # List of all density estimator builders for testing.
