+++ tests/score_estimator_test.py
@@ -9,7 +9,7 @@ import pytest
 import torch
 
 from sbi.neural_nets.embedding_nets import CNNEmbedding
-from sbi.neural_nets.build_nets import build_score_estimator
+from sbi.neural_nets.net_builders import build_score_estimator
 
 
 @pytest.mark.parametrize("sde_type", ["vp", "ve", "subvp"])
